// src/invoices/dto/invoice-preview.dto.ts
export class InvoicePreviewDto {
  cufe: string;
  documentType: string;
  issueDate: string;
  receptionDate: string;
  issuerNit: string;
  issuerName: string;
  receiverNit: string;
  receiverName: string;
  currency: string;
  paymentMethod: string;
  total: number;
  status: string;
  group: string;
}